+++ src/main/groovy/com/dicedmelon/gradle/jacoco/android/ReportConfiguration.groovy
@@ -1,0 +1,18 @@
+package com.dicedmelon.gradle.jacoco.android
+
+public class ReportConfiguration {
+
+  private boolean enabled
+
+  ReportConfiguration(boolean enabled) {
+    this.enabled = enabled
+  }
+
+  public boolean isEnabled() {
+    enabled
+  }
+
+  public void enabled(boolean enabled) {
+    this.enabled = enabled
+  }
+}