+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Make it possible to use the plugin with Java 7
 * Provide default name patterns for classes to be excluded from report generation.
 Currently, these patterns include *Android*, *Dagger 2* and *Butter Knife* generated classes.
+* Allow user to toggle report generation for specified type
+(e.g. `jacocoAndroidUnitTestReport { csv.enabled true }`).
 
 ## Version 0.1.0 *(2015-09-13)*
 * Initial release